feat: improve DeploymentId encoding/decoding
This adds support for decoding/encoding to and from hex strings in
addition to IPFS hashes / CIDv0.

Introducing a dedicated `DeploymentIdError` type makes this a breaking
change, but perhaps we're ok with that given the benefits.

The reason this is useful is that in various places we store deployment
IDs as hex strings (e.g. in the indexer software). Instead of explicitly
handling these cases on one-off basis, the changes introduced here allow
to treat them all the same.